browser.py
from langchain_openai import ChatOpenAI
from browser_use import Agent
from dotenv import load_dotenv
load_dotenv()

import asyncio

llm = ChatOpenAI(model="gpt-4o")

async def main():
    agent = Agent(
        task="Compare the price of gpt-4o and DeepSeek-V3 for 2 million tokens",
        llm=llm,
    )
    result = await agent.run()
    print(result)

asyncio.run(main())
