requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.6.2.post1
asn1crypto==1.5.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bip-utils==2.9.3
bitarray==2.9.2
boto3==1.36.6
botocore==1.36.6
browser-use==0.1.29
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.1
cbor2==5.6.4
cdp-agentkit-core==0.0.11
cdp-langchain==0.0.13
cdp-sdk==0.15.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
cleo==2.1.0
coincurve==20.0.0
courlan==1.3.2
crashtest==0.4.1
crcmod==1.7
cryptography==44.0.0
cytoolz==0.12.3
dateparser==1.2.0
defusedxml==0.7.1
Deprecated==1.2.17
distlib==0.3.9
distro==1.9.0
dulwich==0.22.7
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
fireworks-ai==0.15.12
frozenlist==1.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1rc1
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
hexbytes==1.2.1
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
httpx-ws==0.7.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.7.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.1
keyring==25.6.0
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-aws==0.2.11
langchain-core==0.3.31
langchain-fireworks==0.2.6
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langgraph==0.2.67
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.1.142
lmnr==0.4.55
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
monotonic==1.6
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.2
oauthlib==3.2.2
ollama==0.4.7
openai==1.60.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-langchain==0.37.1
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-threading==0.50b0
opentelemetry-instrumentation-urllib3==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-semantic-conventions-ai==0.4.2
opentelemetry-util-http==0.50b0
orjson==3.10.11
packaging==24.2
parsimonious==0.10.0
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
playwright==1.49.1
poetry==2.0.1
poetry-core==2.0.1
posthog==3.10.0
proto-plus==1.26.0rc1
protobuf==5.29.3
py-sr25519-bindings==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.0.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rlp==4.0.1
rsa==4.9
s3transfer==0.11.2
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
standard-imghdr==3.13.0
tenacity==9.0.0
tiktoken==0.8.0
tld==0.13
tomlkit==0.13.2
toolz==0.12.1
tqdm==4.67.0
trafilatura==2.0.0
trove-classifiers==2025.1.15.22
tweepy==4.15.0
twitter-langchain==0.0.11
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.29.1
web3==7.7.0
websockets==13.0.1
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xattr==1.1.4
yarl==1.11.1
zipp==3.21.0
